chore: impore the typings of the competition, lane, player and score
